requirement.txt
django-environ==0.9.0
et-xmlfile==1.1.0
numpy==1.24.0
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
